Add campaign status and monitor how many keys was added/deleted

+++ contracts/near-campaign/src/clear_state.rs
@@ -10,11 +10,16 @@ pub enum ClearStatus {
 impl Campaign {
   #[private]
   pub fn clear_state(&mut self, keys: Vec<PublicKey>) -> ClearStatus {
+    self.status = CampaignStatus::Deletion;
+
     keys.into_iter().for_each(|pk| {
       let key = pk.into();
-      self.keys.remove(&key);
-      // TODO only if status == 'active' remove access key
-      Promise::new(env::current_account_id()).delete_key(key);
+      self.keys_stats.deleted_during_deletion += 1;
+      let key_status = self.keys.remove(&key);
+
+      if key_status == Some(KeyStatus::Active) {
+        Promise::new(env::current_account_id()).delete_key(key);
+      }
     });
 
     ClearStatus::Completed(self.keys.is_empty())
